[TS migration] Migrate 'withNavigationFocus.js' HOC to TypeScript:
