+++ src/components/RoomNameInput/roomNameInputPropTypes.js
@@ -1,5 +1,4 @@
 import PropTypes from 'prop-types';
-import {withNavigationFocusPropTypes} from '../withNavigationFocus';
 
 const propTypes = {
     /** Callback to execute when the text input is modified correctly */
@@ -29,7 +28,7 @@ const propTypes = {
     /** Whether we should wait before focusing the TextInput, useful when using transitions on Android */
     shouldDelayFocus: PropTypes.bool,
 
-    ...withNavigationFocusPropTypes,
+    isFocused: PropTypes.bool.isRequired,
 };
 
 const defaultProps = {
